Move analysis job actions to root actions file
